Split FacebookServerSideEvent into buffer, sender, and pixel-event queue
FacebookServerSideEvent mixed three responsibilities. Break them apart into
focused classes, all reached only through Signals.

- ServerEventBuffer: the generic request event buffer (record,
  queue_for_send, get_tracked_events, get_num_tracked_events,
  get_pending_events) accumulated and flushed together later.
- PendingPixelEvents: the callback-keyed queue (set/get by callback name) of
  server events held until a specific WordPress callback renders them as pixel
  code (e.g. WooCommerce's add-to-cart fragment).
- ServerEventSender: sends events to the Conversions API and owns the API
  client instance (send, is_open_bridge_event).
- Signals holds all three as shared statics and owns the cross-cutting
  consent gating (should_suppress_frontend_send), the
  before_conversions_api_event_sent filter, and the held -> browser-release
  routing.
- FacebookServerSideEvent deleted.

Also fix stale doc references to types that no longer exist: OpenBridge
("ServerEvent" -> "server event") and ReleaseSignalsAjax ("ServerEventFactory"
-> "EventFactory/UserPiiResolver").

Tests: FacebookServerSideEventTest split into ServerEventBufferTest,
ServerEventSenderTest, and PendingPixelEventsTest; SignalsTest gains the
moved-in filter/suppress/routing cases; drop the stale FacebookServerSideEvent
import from OpenBridgeTest.

Full suite: 196 green; phpcs whole-tree clean.
