--- __tests__/core/FacebookServerSideEventTest.php
@@ -1,337 +1,0 @@
-<?php
-/**
- * Facebook Pixel Plugin FacebookServerSideEventTest class.
- *
- * This file contains the main logic for FacebookServerSideEventTest.
- *
- * @package FacebookPixelPlugin
- */
-
-/**
- * Define FacebookServerSideEventTest class.
- *
- * @return void
- */
-
-/**
-* Copyright (C) 2017-present, Meta, Inc.
-*
-* This program is free software; you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation; version 2 of the License.
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU General Public License for more details.
-*/
-
-namespace FacebookPixelPlugin\Tests\Core;
-
-use FacebookPixelPlugin\Core\EventFactory;
-use FacebookPixelPlugin\Core\FacebookSignalState;
-use FacebookPixelPlugin\Core\FacebookServerSideEvent;
-use FacebookPixelPlugin\Core\FacebookPluginConfig;
-use FacebookPixelPlugin\Tests\FacebookWordpressTestBase;
-
-/**
- * FacebookServerSideEventTest class.
- *
- * @runTestsInSeparateProcesses
- * @preserveGlobalState disabled
- *
- * All tests in this test class should be run in separate PHP process to
- * make sure tests are isolated.
- * Stop preserving global state from the parent process.
- */
-final class FacebookServerSideEventTest extends FacebookWordpressTestBase {
-  /**
-   * A fresh server-event store for each test.
-   *
-   * @var FacebookServerSideEvent
-   */
-  private $store;
-
-  public function setUp(): void {
-    parent::setUp();
-    $this->store = new FacebookServerSideEvent();
-  }
-
-  /**
-   * Tests that the track method records the event.
-   *
-   * This test ensures that when a 'Lead' event
-   * is tracked using the track method,
-   * the number of tracked events increases by one. It verifies that the
-   * get_num_tracked_events method returns the
-   * expected count of 1 after tracking
-   * the event.
-   *
-   * @return void
-   */
-  public function testTrackRecordsEvent() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-
-    $event = EventFactory::create( 'Lead' );
-
-    $this->store->track( $event );
-
-    $this->assertEquals(
-      1,
-      $this->store->get_num_tracked_events()
-    );
-  }
-
-  /**
-   * Tests that the 'before_conversions_api_event_sent' filter is invoked.
-   *
-   * This test verifies that the 'before_conversions_api_event_sent' filter is
-   * applied to the events array before sending the
-   * events using the send method
-   * of FacebookServerSideEvent. It ensures that
-   * the filter modifies the events
-   * as expected.
-   *
-   * @return void
-   */
-  public function testSendInvokesFilter() {
-    $events = array();
-    \WP_Mock::expectFilter( 'before_conversions_api_event_sent', $events );
-
-      $events = $this->store->send( $events );
-  }
-
-  /**
-   * Tests that the track method of FacebookServerSideEvent stores the events
-   * correctly.
-   *
-   * This test verifies that when tracking events using the track method, they
-   * accumulate in order in the pending queue (dispatched together on flush)
-   * and can be retrieved using the get_pending_events method.
-   */
-  public function testStoresPendingEvents() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-
-    $event1 = EventFactory::create( 'Lead' );
-    $event2 = EventFactory::create( 'AddToCart' );
-
-    $this->store->track( $event1 );
-    $this->store->track( $event2 );
-
-    $pending_events =
-    $this->store->get_pending_events();
-
-    $this->assertEquals(
-      2,
-      $this->store->get_num_tracked_events()
-    );
-    $this->assertEquals(
-      2,
-      count( $pending_events )
-    );
-    $this->assertEquals(
-      'Lead',
-      $pending_events[0]->getEventName()
-    );
-    $this->assertEquals(
-      'AddToCart',
-      $pending_events[1]->getEventName()
-    );
-  }
-
-  /**
-   * Tests that the set_pending_pixel_event and
-   * get_pending_pixel_event methods of FacebookServerSideEvent
-   * store the events correctly.
-   *
-   * This test verifies that when setting a pending pixel
-   * event using the set_pending_pixel_event method,
-   * the events are stored in the correct order.
-   * It also verifies that the events can be retrieved using the
-   * get_pending_pixel_event method using the correct callback name.
-   */
-  public function testStoresPendingPixelEvents() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-
-    $event = EventFactory::create( 'Lead' );
-
-    $this->store
-      ->set_pending_pixel_event( 'test_callback', $event );
-
-    $pending_pixel_event = $this->store
-      ->get_pending_pixel_event( 'test_callback' );
-
-    $this->assertEquals(
-      'Lead',
-      $pending_pixel_event->getEventName()
-    );
-  }
-
-  /**
-   * Tests that frontend sends are suppressed while signals are held.
-   *
-   * @return void
-   */
-  public function testSendSuppressedWhenHeldOnFrontend() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-
-    \WP_Mock::userFunction( 'is_admin', array( 'return' => false ) );
-    \WP_Mock::userFunction( 'wp_doing_cron', array( 'return' => false ) );
-
-    FacebookSignalState::hold();
-
-    $api = \Mockery::mock( 'alias:FacebookPixelPlugin\FacebookAds\Api' );
-    $api->shouldReceive( 'init' )->never();
-
-    $event  = EventFactory::create( 'Lead' );
-    $result = $this->store->send( array( $event ) );
-
-    $this->assertNull( $result );
-  }
-
-  /**
-   * Tests that send() returns error when circuit breaker is open.
-   */
-  public function testSendReturnsErrorWhenCircuitOpen() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-    \WP_Mock::userFunction( 'is_admin', array( 'return' => false ) );
-    \WP_Mock::userFunction( 'wp_doing_cron', array( 'return' => false ) );
-
-    \WP_Mock::userFunction(
-      'get_transient',
-      array(
-        'args'   => array(
-          FacebookPluginConfig::CONNECTION_INVALID_TRANSIENT,
-        ),
-        'return' => time() - 60,
-      )
-    );
-
-    $api = \Mockery::mock( 'alias:FacebookPixelPlugin\FacebookAds\Api' );
-    $api->shouldReceive( 'init' )->never();
-
-    $event  = EventFactory::create( 'Lead' );
-    $result = $this->store->send( array( $event ) );
-
-    $this->assertFalse( $result['success'] );
-    $this->assertStringContainsString( 'circuit', $result['error']['message'] );
-  }
-
-  /**
-   * Tests that send() blocks with a circuit-open error only when the
-   * transient is fresh, not when it's stale (half-open).
-   */
-  public function testSendBlocksOnlyWhenTransientFresh() {
-    self::mockFacebookWordpressOptions();
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-    \WP_Mock::userFunction( 'is_admin', array( 'return' => false ) );
-    \WP_Mock::userFunction( 'wp_doing_cron', array( 'return' => false ) );
-
-    // Fresh transient → circuit open → should block.
-    \WP_Mock::userFunction(
-      'get_transient',
-      array(
-        'args'   => array(
-          FacebookPluginConfig::CONNECTION_INVALID_TRANSIENT,
-        ),
-        'return' => time() - 10,
-      )
-    );
-
-    $event  = EventFactory::create( 'Lead' );
-    $result = $this->store->send( array( $event ) );
-
-    $this->assertFalse( $result['success'] );
-    $this->assertStringContainsString(
-      'circuit',
-      $result['error']['message']
-    );
-  }
-
-  /**
-   * Tests that send() returns null when pixel_id or access_token is empty.
-   */
-  public function testSendReturnsNullWhenNoCredentials() {
-    self::mockFacebookWordpressOptions(
-      array(
-        'pixel_id'     => '',
-        'access_token' => '',
-      )
-    );
-
-    \WP_Mock::userFunction(
-      'sanitize_text_field',
-      array(
-        'args'   => array( \Mockery::any() ),
-        'return' => function ( $input ) {
-          return $input;
-        },
-      )
-    );
-    \WP_Mock::userFunction( 'is_admin', array( 'return' => false ) );
-    \WP_Mock::userFunction( 'wp_doing_cron', array( 'return' => false ) );
-
-    $event  = EventFactory::create( 'Lead' );
-    $result = $this->store->send( array( $event ) );
-
-    $this->assertNull( $result );
-  }
-}
